requirements.txt
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
imbalanced-learn==0.8.1
matplotlib==3.5.3
seaborn==0.11.2
shap==0.40.0
lime==0.2.0.1
pyyaml==5.4.1
transformers==4.20.1
torch==1.10.2
